OS detection framework
This patch adds a function that determines the current operating system.
Right now we only detect Arch Linux.  I've also added an `info`
subcommand (which is the default if you don't explicitly name a
subcommand), which prints out the location of the package database and
the OS that was detected.