Better OS detection
This is a lot cleaner.  The detection function sets the
`buzzy.config.os` global variable if it detects an operation system.
This variable is now a Python object, which we will expect to implement
a particular interface.  The OS objects can still define extra
subcommands, but these are included right in the OS object, rather than
us playing weird module-loading games to find them.