Installing native packages on Arch
You can now mark a package has being natively available on Arch Linux by
adding the following to your package definition:

    arch:
        native: [package-name]

This patch also includes new helper functions for calling subprocesses,
with or without sudo.  Under normal operation, we'll collect the output
of the subprocess, and only display it if the subprocess fails.  By
giving a `--verbose` option, we'll go ahead and print out all of the
output as it happens.