Persistent options moved to an environment
We now store persistent options in a YAML configuration file in an
“environment” directory.  This is also where we will (by default) build
any packages.  It doesn't have to be (and probably shouldn't be) inside
of your recipe database; one of the configuration options is the
location of the database.

End result, you should run all of the buzzy commands from within an
environment directory.  And you should use the `buzzy configure` command
to set up your environment.