`buzzy build` command
This patch adds a command that builds a package, but doesn't install it.
We make sure that any dependencies and build dependencies are built and
installed.

This also adds a `--force` option to the `build` and `install` commands,
which causes the packages listed on the command line to be build and
installed even if they're already built or installed.  The `--force-all`
option does this for all of the package's dependencies, too.