`buzzy update` command
This patch adds the `buzzy update` command, which builds and installs
every recipe in the recipe database.  If you provide the `--force`
option, we'll build and install each package, even if it's already built
and/or installed.  Otherwise we'll only build and install the packages
that aren't up to date.