Python packages
We can now build Python packages on Arch.  We produce two packages for
each recipe: one for Python 2, and one for Python 3.  For any
dependencies that are also Python recipes, we only require the package
that corresponds to the same Python version.  If a Python recipe depends
on a non-Python recipe, whatever packages marked as the "default" is
used.  If a non-Python recipe depends on a Python recipe, we use the
Python 3 package as the default dependency.

This patch also moves a lot of the high-level interface for creating
packages from recipes, and the tag mechanism for resolving these kinds
of dependencies, into the generic recipe/package code.  The
Arch-specific modules now really do only contain Arch-specific logic.