Buzzy version strings
Since we're going to support a bunch of different packaging systems,
each of which has a slightly different versioning setup, we're going to
have a bit of a mess on our hands.

This patch adds a type for representing version numbers.  I'm trying to
make it relatively easy to parse version numbers from a bunch of
different schemes, and still have them be interchangeable.  We'll see if
that works.

In this setup, a version is a list of parts.  Each part is either a
release tag, a pre-release tag, or a post-release tag.  (Each version
also ends with a special "final" tag, which is only there to make the
comparison algorithm work.)  This is very much like how setuptools
parses version strings, but we're not going to require setuptool's
particular encoding of versions as strings.

With this representation, we can compare versions lexicographically,
where individual parts are ordered as follows:

  pre < final < post < release

This works perfectly, as long as we make sure to not include any "0"
release parts in the comparison, unless they're followed by a non-"0"
release tag.  (This is to ensure that 2.4+dev compares greater than
2.4.0, but less than 2.4.1, for instance.)

We do have a string encoding for these version numbers, where each part
is explicitly prefixed by a ".", "~", or "+" to indicate whether it's a
release, pre-release, or post-release tag.