Package dependencies
A package dependency is the request to perform some operation on some
package.  The goal of Buzzy will be to perform actions that satisfy
these dependencies.  The most common operation is `install`, which
causes the package to be installed onto the local system in some way
that makes it usable by the user, or by other packages.  Other
operations might include `build` or `test`, which would be useful during
development.