Execute subprocesses
This patch adds some helper functions for executing subprocesses.  There
are a handful of use cases:

  • Just run the command.  Silently drop the output unless the user has
    requested verbosity; in that case, allow the output to go directly
    to the real stdout or stderr.

  • Capture the output.

  • Ignore the output, and test whether the command executed
    successfully.

For testing purposes, you have the ability to mock out each subprocess
that's called, by giving an explicit exit code, and stdout and stderr
stream; in that case, we'll act like that's what occurred when the
command is executed.  You can also allow a particular command to
actually be executed; you should only do that if you've verified that
the command is actually available, and its output and exit code are
completely reproducible.