Multiple ways to specify subprocess command
This patch adds several variants of each of our functions for executing
a command in a subprocess.  The variants differ in how the name of the
executable, and the parameters to the command, are specified — as a
`cork_array`, a `va_args`, or a variable-length parameter list.  In all
cases, you must pass in an array of strings, where the first string is
the name of the executable.