Arch version strings
This patch adds functions that convert Buzzy versions to and from the
equivalent Arch Linux version strings.  (I.e., an Arch version string
that looks as identical as possible, but compares the same way.)  Arch
treats an alphabetic suffix as a prerelease if there's no intervening
`.`, and as a postrelease if there is one.  This gives the following
Buzzy → Arch conversions:

   2.0           2.0
   2.0~alpha1    2.0alpha1
   2.0~1         2.0pre1
   2.0+hotfix1   2.0.hotfix1
   2.0+1         2.0.post1