Detect Arch Linux
We detect whether the current machine is running Arch Linux by looking
for the /etc/arch-release file.  If it exists, we assume that we're
running on an Arch system.  We don't care about the contents of the
file.  (In fact, since Arch is a rolling release, the file is usually
empty.)