Actions
The `bz_action` and `bz_action_phase` types are how we're going to
manage the sequence of actions that need to be performed while building
or installing a collection of packages.  Each action instance includes a
function that will be called to perform the action, and a separate
function to determine whether the action needs to be performed.  Actions
can also define prerequisites and postrequisites, which allow you to
locally define an ordering on actions.

Like subprocesses, actions can be mocked; in this case, instead of
printing out the list of actions that we're performing, we append the
list of actions into a buffer.  Our test cases can then verify which
actions were performed, and the specific text that would be outputted
describing those actions.