Cached package databases
This patch adds a helper package database implementation that
automatically caches packages for you.  You provide a satsify method
that just blindly creates a package for a dependency.  The cache helper
translates that into a satisfy method that will check if we've already
created a package for a particular dependency, and if so, return it.
This will help ensure that if two packages have a dependency on a third,
we'll automatically only include the install action for the third
package once.