Pacman package creation
This patch adds support for creating binary packages using Arch Linux's
pacman packager.  This includes a new “package spec” object, which will
be used throughout Buzzy to hold various details about a new package
being created.

This patch also incorporates some of libcork's new file and directory
functions.  Most important among them is the ability to find the
user-specific cache directory (defaults to ~/.cache, but can be
overridden via the XDG_CACHE_HOME environment variable), which is where
we're going to put our build artefacts.  (In the old Python version of
Buzzy, we put things in a subdirectory of `/tmp`, but this seems to be
more in the spirit of the current directory specifications.)