CMake builder
This patch adds an interface for defining “recipes”, which are
instructions for building a package from source.  It also provides an
implementation of this interface for building projects that use CMake.
There are test cases that verify that the expected commands are
executed, but this isn't linked into the `buzzy` command-line program
yet.