Storing environment variables in a hash table
Most of the time, we're going to build up a set of environment variables
in memory, storing them in a simple hash table.  This patch adds a class
for managing these kinds of variable sets — it's basically a wrapper
around a `cork_hash_table` with functions for adding and retrieving the
value providers for a particular key.