Environment value sets
This patch adds an interface for retrieving an environment variable from
a set of variables, regardless of how that set is stored and queried.
(Eventually we'll add an implementation that reads values from a YAML
file lazily, for instance.)  The patch also includes an implementation
of this interface that gets values from an instance of the hash table
class from the previous patch.