Environments
An environment is a mapping of variable names to values, implemented
using all of the types introduced in previous patches in this series.
An environment contains several value sets.  These value sets are
organized into two groups: “sets” and “backup sets”.  When you query an
environment for a variable, all of the sets are checked first, in the
order that they were added to the environment.  Then, all of the backup
sets are checked.  (This allows us to mark a value set as provide
default values, for instance, without having to worry about making sure
it's added to the environment last.)