Global and package-specific environments
This patch adds an singleton “global” environment, which will contain
environment variables that apply to every package we know about.  This
will eventually be populated via the command line and configuration
files.  The patch also adds a constructor function for creating
package-specific environments, which will contain furthers variables
that only apply to a single package.

The main feature that these new functions add over the existing
environment constructor is that we maintain a list of default values for
predefined variables.  The global environment, and each package-specific
environment, will automatically contain a variable set that retrieves
values from this set of defaults, if a variable isn't given an explicit
value in the environment in question.

Right now, all of the default values are stored in a single variable
set.  At some point, we might need to separately store the default
values for those variables that should only appear in package-specific
environments.  But for now, we don't need to do that.