buzzy raw build
This patch adds the `buzzy raw build` command, which builds a package
whose source code is in the current directory.  Like the `buzzy raw pkg`
command, you have to provide all of the necessary package configuration
variables explicitly on the command line, including the `builder`
option, which selects which builder implementation is used to actually
perform the build.