Refactor built package interface
This patch consolidates a couple of different parts of the API that are
used to build packages from source.  There are now three orthogonal
interfaces that together tell Buzzy how to build and install a package
from source: an “unpacker” (which doesn't exist yet), a “builder”, and a
“packager”.  There's also now a standardized design for the builder and
packager — they're basically just a collection of two or three actions
that perform the real work.

There are currently two available builders (noop and cmake), and two
available packagers (noop and pacman).  We autodetect which packager to
use, based on the distribution that the current machine is running.
Eventually we'll also autodetect a builder by inspecting the package's
source tree.