Consistent results with non-forking test cases
The check framework usually runs each test case in a separate process,
so that it can test for segfaults without affecting the harness process.
You can override this behavior with the `CK_FORK` environment variable.
We do that sometimes during our valgrind tests to reduce the number of
crufty false positives.  But, we were being too aggressive with some
caching, which meant that the expected test case results didn't line up
if we weren't starting with a fresh empty process each time.  This patch
removes some of that caching (since it's not like the underlying
operations are very expensive), so that we can have more reproducible
test cases.