Use strings in file manipulation functions
Before we were constructing `cork_path` objects by hand before passing
them in to the various file manipulation functions.  But now we use
interpolated values for everything, and it's easier to just grab the
value as a string and pass that in.  Let the helper functions worry
about creating the `cork_path` and `cork_file` objects when they're
needed.