Base paths for value sets
You can now provide a “base path” for each of the value sets in an
environment.  If you use `bz_env_get_path` to retrieve the value of an
environment variable, and the value set that provides the value has a
base path set, and the content of the variable is a relative path, then
we'll automatically concat the two together before returning the result.
We're going to use this in our filesystem repositories, so that we can
correctly interpret any relative paths mentioned in a YAML configuration
file.

(Note that a base path is a property of a value set, not of the
environment as a whole.  That's because we might have environments that
consist of several value sets, each of which is defined in a different
place, and which therefore should have a different base path.)