YAML environments
This patch adds functions for loading in an environment value set from a
YAML file.  The YAML file must contain a top-level mapping node.  We
currently only support looking for bare string keys in that top-level
mapping.  When we find a value in the YAML file, it must currently be a
string.  We assume that the string is an interpolated value, so you can
put ${var} references in the YAML content.  We might eventually support
additional kinds of YAML values (probably based on tags, which would
then be used to instantiate different value providers).  Soon we will
also support nested mappings in the YAML, with environment variables
with names like `a.b.c` being used to delve deeper in to the YAML
content.