Cache values in YAML environments
The YAML value set implementation now caches any variables that you
retrieve from the YAML file.  This gives us a slight (but probably not
noticeable) speed boost when we read the same value from the YAML file
multiple times.  More importantly, it means that we don't have a memory
leak, since the value set is responsible for freeing any values that it
returns.  (Before, we weren't keeping track of the value instances at
all, so we never had a chance to free them.)