Hide phases with no actions
The `bz_action_phase_perform` function now takes in a flags parameter
that lets you control how it works.  The new `BZ_ACTION_HIDE_NOOP` flag
will cause us to not print out any messages if there aren't any actions
to perform in the phase.  Without this flag, we'll print out the header
message and a "Nothing to do" message.