Repository registry
The function that loads in a filesystem repository from the current
directory now add the repository to our new repository registry.  The
`buzzy info` command now prints out the location of each repository in
the registry.  This is how we'll support linked repositories in the
future.

The `buzzy doc` command now loads in the repository registry, too, and
uses the “main” (i.e., current directory) repository, if there is one,
when printing the value of an environment variable.