Refactor loading from a filesystem repository
This patchs moves all of the functions that actually load the contents
of a filesystem repository into the repo's load method.  (Before we were
performing some of the steps in the repo constructor.)  The loading
logic has also been moved into a separate helper function.  This means
that was used to be called a “filesystem” repo is now a “local
filesystem” repo — i.e., one where the repository directory is already
present on the local filesystem.  Repositories that are cloned from some
remote source will be completely separate repo implementations; they'll
just all happen to have a call to this helper function somewhere in
their `load` methods.