The great variable renaming
This patch changes the name of just about every environment variable.
I'm trying to be more consistent with how directories are named, and how
variables are nested into nice namespaces.  So now, all of the variables
that have to do with a repository start with `repo.`.  Variables for
identifying directories now always end in `_dir`, and variables
identifying files end in something specific to the type of file (e.g.,
`_yaml`).