Remove action type
Actions are a thing of the past.  Instead of having a separate type to
manage actions, we just do stuff directly.  An “action” is now just a
regular C function; when it returns, the action should have been
performed.  If it has prereqs, it should recursively call the functions
that perform the prereq.  If we should only perform the action once, it
should somehow keep track of whether it's been executed already.

The main thing we were using actions and action phases for were to keep
track of the number of actions that would eventually be run, so that we
could have a nice `[XX/YY]` prefix when we logged which actions were
being performed.  But it complicates things, especially since we're
going to have remote repositories, whose contents will be loaded in an
action.  If those remote repos have dependencies on other repos, we
won't know that until we've started the action that clones and loads its
contents.  That messes up counts anyway, so for now, we're going to
abandon the total action count, and go with the much simpler API.