git repositories
We now have the ability to clone a Buzzy repository from a remote git
repository.  Right now this functionality is only available in the C
API; next step is to add it to the URL and YAML repository handlers so
that you can clone a git/Buzzy repo via a repository link.