buzzy uninstall
The new `buzzy uninstall` command removes a package from the system.  We
don't check the version that was installed; if some other version that
the “current” version is installed, we'll go ahead and remove that.  We
also don't check any forward dependencies; most package managers will
refuse to uninstall a package if there's some other package installed
that depends on it.