New value interface
I've extracted out the part of the old `bz_env` type that stored values
for variables.  This now belongs to a separate `bz_value` type.  The old
“value provider” and “value set” types are now scalar values and map
values, respectively, rather than being distinct types.  I've also added
support for array values.  This results in some pretty sweeping changes
in the implementation of the environment class, but everything
downstream of that is more or less unchanged.  (The most invasive change
is that when you grab the value of a variable, you are never responsible
for freeing the result.)