Dependencies
You can now add a `dependencies` and/or `build_dependencies` variable to
your package descriptions.  If given, buzzy will ensure that those
dependencies are present when building and installing the package.  Our
pacman packager also uses the `dependencies` variable to define the
dependencies of the binary package that it produces, so that the pacman
database correctly encodes this information.