Set base path for repo and package environments
So that relative paths are handled correctly.  For repositories, the
base path is easy; it's whatever we set for the `repo.base_dir`
variable.  Right now we only create new package environments for
single-package filesystem repositories.  For those, the base path of the
package is the value of the repository's `repo.base_dir` variable.