CentOS and Ubuntu Vagrant boxes
This patch adds a Vagrant configuration that defines two VMs: one
running CentOS 6.4, and one running Ubuntu 12.10.  If you run `vagrant
up` from within the new `vagrant` directory, it should create both VMs
and install all of the necessary build tools into them.  You can also
run:

  $ vagrant ssh [VM name] -c /vagrant/vagrant/build

to build and test all of the code.  This needs to be automated a bit
more, but it's a good start.