Don't use `cmake --target`
Our cmake builder now assumes that we'll end up with Makefiles in the
build directory, and runs `make` directly to build, test, and stage the
package.  Before, we were using cmake's `--target` option to let cmake
decide which command to run to perform the actual build.  But this
option wasn't added until cmake 2.8, which isn't (easily) available on
some of the platforms we want to support.