Configuration files that override global defaults
You can now create a handful of YAML configuration files, whose contents
will override any precompiled default values.  The following files are
checked, with earlier files taking precedence over later files:

  $XDG_CONFIG_HOME/buzzy.yaml
    (which is usually $HOME/.config/buzzy.yaml)
  $XDG_CONFIG_DIRS/buzzy.yaml
    (which is usually /etc/xdg/buzzy.yaml)
  $XDG_DATA_HOME/buzzy/config.yaml
    (which is usually $HOME/.local/buzzy/config.yaml)
  $XDG_DATA_DIRS/buzzy/config.yaml
    (which is usually $PREFIX/share/buzzy/config.yaml)

The final one is intended for default values that would be cumbersome
to compile in to the buzzy executable.  We don't have any of those yet,
so that file doesn't exist.  But if we want to use this mechanism to
define default values, we'll just need to create a new config.yaml file
and have the cmake build scripts install it into $PREFIX/share/buzzy.