Access variables in linked repositories
If a “child” repository contains a link to a “parent” repository, you
can now access any of the parent's variables from within the child
repository.  This allows us to extract commonly used variables (such as
a set of native-package overrides) into a single repository, and link to
that repository to use it in a bunch of package definitions.

Note that we can't provide the reverse (allowing the parent to access
variables in the child), since we don't want the contents of the parent
repo to vary depending on which child links to it.  (Note, however, that
if an variable in the parent has a ${var} reference, that reference it
looked up starting from the child repo, which allows parents to define
specific “hooks” that can be overridden in children.)