Autotools builder
This patch adds a basic autotools builder.  We can automatically detect
packages that use the autotools, by looking for a `configure` or
`configure.ac` file.  (The names of these files can be overridden — for
instance, if the autoconf file is called `configure.in`.)  We'll run
`autoreconf` to generate the `configure` script if it doesn't exist.
Then we'll run `configure` and `make` to actually build the package.