Additional installation prefix variables
The `install_prefix` variable has been replaced with a handful of new
variables, which allow you to further customize the locations for each
kind of artefact that a built package installs.  The CMake and autotools
builders have been updated to pass the values of those variables into
the corresponding configuration programs.