Install native Homebrew packages
Buzzy can now find and install native packages from Homebrew.  We parse
the results of `brew info [package]` to determine which versions are
available and/or installed, and `brew install` and `brew remove` to
install and uninstall a package.  There's currently no support for
custom taps or anything other fancy Homebrew features.