Set PKG_CONFIG_PATH during Homebrew builds
The autotools and cmake builders now set the `PKG_CONFIG_PATH`
environment variable during the configure step, if the corresponding
`pkgconfig.path` Buzzy variable is set.  On Mac/Homebrew, we
automatically set this Buzzy variable to point into the $PREFIX/opt
directory of each dependency.  Some of these directories might not
exist, but pkgconfig will silently ignore those, so that's not a
problem.