Don't require check test harness
We use check [1] as one of our testing harnesses.  RHEL5 doesn't provide
a new enough version of check for our purposes, though.  We can build
and install a newer version using Buzzy, but not until Buzzy is built:
a dastardly Catch-22.  So now we make all of the buzzy tests optional.
This lets us bootstrap and install a version of Buzzy, which we can then
use to build and install a newer version of check, which we can then use
to go back and run the Buzzy tests.

[1] http://check.sourceforge.net/