Recursive pkgconfig.path on Mac/Homebrew
Before, we would set the pkgconfig.path variable on Mac/Homebrew to look
for .pc files in the package-specific directories in the Homebrew
prefix's opt directory.  This ensures that we can find any keg-only
dependencies.  Turns out this is not enough; we need to do that
recursively for any dependency reachable from the current package, not
just for its direct dependencies.