Pre- and post-installation scripts
You can now set the `pre_install_script` and `post_install_script`
package variables to tell the package manager to run a script just
before or after a package is installed.  Like all other path variables,
relative paths are interpreted relative to the package's source
directory.  (You can prepend `${build_dir}` to make a path that's
relative to the build directory, in case you have to generate the script
as part of the build process.)