Call ldconfig in RPM and Debian packages
The RPM and Debian package installers don't automatically call ldconfig
for you if your package contains shared libraries.  So when building
packages for those systems, we need to include an explicit call to
ldconfig in the post-install and post-remove scripts.  Ideally we'd
check whether the package contains any shared libraries, and only add in
the call if necessary, but we don't currently have that level of
introspection available.  So we call it for all buzzy-built packages.

Note that pacman *does* call ldconfig for us, so we don't need to add
any explicit calls on that platform.