Relocatable RPMs
You can use the new `relocatable` package property to indicate that the
contents of a built package are relocatable.  The RPM packager will
check this property and build a relocatable RPM if this property is set.
Note that we don't verify that the package contents really are
relocatable — it's up to you not to lie!