Building using the Docker images
We have Docker images for a bunch of different Linux distros, so this
patch adds a helper script for running buzzy inside of one of those
images to build packages for a different distro than your native one.
As an example:

    $ buzzydock centos:centos7 install

The packages will end up in $HOME/.cache/buzzy/distro/centos-centos7.